[MOV] account, account_payment_term: merge module
We added the account_payment_term module in stable, this pr will do some
cleaning by merging the module in account.

task: 3609320
