+++ addons/account/data/account_data.xml
@@ -97,6 +97,14 @@
                 Command.create({'value': 'percent', 'value_amount': 100.0, 'nb_days': 30})]"/>
         </record>
 
+        <record id="account_payment_term_90days_on_the_10th" model="account.payment.term">
+            <field name="name">90 days, on the 10th</field>
+            <field name="note">Payment terms: 90 days, on the 10th</field>
+            <field name="line_ids" eval="[
+                Command.clear(),
+                Command.create({'value': 'percent', 'value_amount': 100.0, 'delay_type': 'days_end_of_month_on_the','nb_days': 90, 'days_next_month': 10})]"/>
+        </record>
+
         <!-- Account-related subtypes for messaging / Chatter -->
         <record id="mt_invoice_validated" model="mail.message.subtype">
             <field name="name">Validated</field>
