+++ addons/account/views/account_payment_term_views.xml
@@ -68,6 +68,8 @@
                                         <field name="value" nolabel="1"/>
                                         <field name="nb_days" string="After"/>
                                         <field name="delay_type" nolabel="1"/>
+                                        <field name="display_days_next_month" column_invisible="1"/>
+                                        <field name="days_next_month" width="30px" nolabel="1" invisible="not display_days_next_month"/>
                                     </tree>
                                 </field>
                             </group>
