--- addons/account_payment_term/models/__init__.py
@@ -1,3 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import account_payment_term
