--- addons/account_payment_term/views/account_payment_term_views.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-        <record id="account_payment_term_form" model="ir.ui.view">
-            <field name="model">account.payment.term</field>
-            <field name="name">account.payment.term.form</field>
-            <field name="inherit_id" ref="account.view_payment_term_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="//field[@name='line_ids']/tree['delay_type']" position="inside">
-                    <field name="display_days_next_month" column_invisible="1"/>
-                    <field name="days_next_month" width="30px" nolabel="1" invisible="not display_days_next_month"/>
-                </xpath>
-            </field>
-        </record>
-    </data>
-</odoo>
